+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsDiagnosticTestGenerated.java
@@ -580,6 +580,12 @@ public class JsDiagnosticTestGenerated extends AbstractJsDiagnosticTest {
       }
 
       @Test
+      @TestMetadata("GraphExtensionFactory_WrongAnnotation.kt")
+      public void testGraphExtensionFactory_WrongAnnotation() {
+        run("GraphExtensionFactory_WrongAnnotation.kt");
+      }
+
+      @Test
       @TestMetadata("GraphPrivateBindingNotVisibleToChild.kt")
       public void testGraphPrivateBindingNotVisibleToChild() {
         run("GraphPrivateBindingNotVisibleToChild.kt");