+++ CHANGELOG.md
@@ -6,12 +6,19 @@ Changelog
 
 ### Fixes
 
+- **[FIR]** Report a diagnostic error when `@DependencyGraph.Factory` is used for a graph extension, instead of crashing.
 - **[IR]** Fix `createGraphFactory()` calls for graph factory interfaces compiled in upstream modules with IR-only class generation.
 - **[IR/interop]** Fix a crash when member-injecting an unscoped constructor-injected binding as `dagger.Lazy<T>`.
 
 ### Changes
 
 - Fold `2.4.20-dev-3583` compat into `2.4.20-dev-6138` as IntelliJ 2026.2 RCs now depend on a newer version.
+
+### Contributors
+
+Special thanks to the following contributors for contributing to this release!
+
+- [@bnorm](https://github.com/bnorm)
 
 ### [Consider sponsoring Metro's development](https://www.zacsweers.dev/sponsoring-metro/)